ITMSPDiscovery/requirements.txt
google-search-results>=2.4.2
httpx>=0.27.0
python-dotenv>=1.0.0
google-genai>=1.0.0
requests>=2.31.0
